ANDROID: GKI: Enable CONFIG_USB_TYPEC_UCSI
GKI 1 had CONFIG_USB_TYPEC and CONFIG_USB_TYPEC_TCPM enabled, but forgot
to enable the _UCSI driver as well.  Do that now as everyone relies on
this class driver.

If platform-specific TYPEC changes are needed, just enable the
platform-specific tyepc driver for the specific platform.

Note, for some reason x86 forgot to enable TYPEC entirely, so fix that
up and enable it here so we catch build issues sooner.

Bug: 157965270
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I353eec6ed4f73525408033f283081b8013ede874
Signed-off-by: Matthias Maennich <maennich@google.com>