Revert "security,lockdown,selinux: implement SELinux lockdown"
This reverts commit f1ee68a8f6bdddcc93b9a582bf49c3113ff4f180.

Reason for revert:
The change being reverted adds a new "lockdown" audit class. Support
for this new class needs to be added to Android and the processes
which need to be part of this class have to be annotated. While support
for this class has not yet been added to Android, this lockdown class
will be removed.

Tracefs usage by Android triggers a violation with respect to this new
audit class which prompted the need for this patch.

Bug: 148822198
Change-Id: Ie06f4be699234fb671ec4bcfe11962b2055a0c60
Signed-off-by: Ram Muthiah <rammuthiah@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>